Implement most of the Lazy Code Motion algorithm
Things added:
- Class for expressions (Expression.cpp)
- General framework for Dataflow Analysis using worklist algorithm (DataflowAnalyzer.h)
- In LazyCodeMotion.cpp, the following is implemented (hopefully correctly):
  - Pre-transformations of CFG -> Single Instruction Basic Blocks and New blocks between critical edges
  - Analysis Phase 1: Anticipated Expressions
  - Analysis Phase 2: Will be available Expressions
  - Analysis Phase 3: Postponable Expressions
  - Analysis Phase 4: Used Expressions
  - Computing the Earliest and Latest sets which correcly allow us to find places to insert new expressions.

The only thing remaining is to now add the new expressions at computed places and transform CFG
accordingly. The main challenge is to change the existing instructions when adding a new instruction.
E.g. if we add t = a + b, we need to change future occurences of a+b to t. When adding same instruction in
diff blocks, we also need to add phi blocks and handle cases correctly.